Create components folder with cache as the first one
Helper cache_connector & module cache are merged into one single class
